create the $HOME folder the .cryptex folder in post install; implement the wait for the subprocess opened in encryption mode

+++ pycryptex/main.py
@@ -1,7 +1,3 @@
-# This is a sample Python script.
-
-# Press ⌃R to execute it or replace it with your code.
-# Press Double ⇧ to search everywhere for classes, files, tool windows, actions, and settings.
 import os
 import subprocess
 import sys
@@ -9,7 +5,7 @@ from pathlib import Path
 from pycryptex.crypto import rsa
 import pycryptex
 from os import path
-
+import time
 import click
 import toml
 
@@ -30,7 +26,7 @@ pass_config = click.make_pass_decorator(Config, ensure=True)
 
 @click.group()
 @click.version_option(version=None, message="pycryptex CLI application (version: %(version)s)")
-@click.option('--verbose', is_flag=True, help='bool, to specify if needed a verbose mode')
+@click.option('--verbose', "-v", is_flag=True, help='bool, to specify if needed a verbose mode')
 @pass_config
 def cli(config, verbose):
     """
@@ -52,7 +48,7 @@ def cli(config, verbose):
 @cli.command()
 @click.argument('file', required=True)
 @click.option('--pubkey', default="my_key.pub", help='(optional) specify the RSA public key')
-@click.option('--remove','-r', is_flag=True, help="(optional, bool=False) to indicate if remove or not the file")
+@click.option('--remove', '-r', is_flag=True, help="(optional, bool=False) to indicate if remove or not the file")
 @pass_config
 def encrypt(config, file, pubkey, remove):
     """Encrypt a file"""
@@ -72,8 +68,10 @@ def encrypt(config, file, pubkey, remove):
 @cli.command()
 @click.argument('file', required=True)
 @click.option('--privkey', default="my_key", help='(optional) specify the RSA private key')
-@click.option('--remove', '-r', is_flag=True, help="(optional, bool=False) to indicate if remove or not the encrypted file")
-@click.option('-s', is_flag=True, help="(optional, bool=False) to indicate destroy file after open")
+@click.option('--remove', '-r', is_flag=True, help="(optional, bool=False) passing this option the encrypted file will"
+                                                   "be removed")
+@click.option('-s', is_flag=True, help="(optional, bool=False) passing this option the decrypted file will"
+                                       "be removed")
 @click.option('--pager', '-p', is_flag=True,
               help="(optional, bool=False) to open or not the pager to read decrypted file")
 @pass_config
@@ -86,10 +84,15 @@ def decrypt(config, file, privkey, remove, s, pager):
         f = rsa.decrypt_file(file=file, private_key=privkey, remove=remove)
         # open file in a pager
         if pager:
-            output = subprocess.run([pycryptex.config_file['config']['pager'], f])
-        # if True delete the decrypted file
-        if s:
-            os.remove(f)
+            exit_code = subprocess.call([pycryptex.config_file['config']['pager'], f])
+            if exit_code == 0:
+                # if True delete the decrypted file
+                time.sleep(pycryptex.config_file['config']['wait_delete_time'])
+                if s:
+                    os.remove(f)
+            else:
+                click.echo(click.style(f"Houston, we have a problem: the opened subprocess has a return value equal to"
+                                       f" {exit_code}", fg="red", bold=True))
         if config.verbose:
             click.echo(click.style(f"priv_key used is: {privkey}", fg="magenta", bold=False))
         click.echo(click.style("File decrypted successfully!", fg="green", bold=True))
@@ -114,5 +117,5 @@ def read_config():
                 }
             }
     except Exception as e:
-        print("Houston, we have a problem in read_config: ", e)
+        click.echo(click.style(f"Houston, we have a problem in read_config: {e}", fg="red", bold=True))
         sys.exit(1)
